ceph-volume: dispatch subcommands through bootstrap()
This is a refactor commit that restructures subcommand dispatching. The
goal is to make inter-module calls more comfortable and uniform.
One can now simply pass a namespace object to main() instead of a list of
cli args or use an implementation specific method of a subcommand. This should
also improve unit testing as its easier to circumvent arg parsing.

Signed-off-by: Jan Fajerski <jfajerski@suse.com>